Fixed: Owner of the image is only authorized to delete the image.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -153,7 +153,22 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,
+                                    Model model,
+                                    HttpSession session) {
+        Image currentImage = imageService.getImage(imageId);
+        List<Tag> tags = currentImage.getTags();
+        model.addAttribute("tags", tags);
+        model.addAttribute("image", currentImage);
+
+        //If current image dont belongs to logged in user
+        User loggeduser = (User) session.getAttribute("loggeduser");
+        if (!currentImage.getUser().getId().equals(loggeduser.getId())) {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
